Curency converting: viewed tour + bill

+++ src/main/java/com/traveloveapi/DTO/service_package/OptionDTO.java
@@ -11,6 +11,6 @@ public class OptionDTO {
     private int group_number;
     private int option_number;
     private String name;
-    private float price;
-    private float price_special;
+    private Double price;
+    private Double price_special;
 }
